experiments/arena/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
about-time==4.2.1
absl-py==2.2.2
alive-progress==3.2.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6 ; sys_platform == 'win32'
cramjam==2.9.1
cryptography==44.0.2
deepdiff==6.7.1
docstring-parser==0.16
fastparquet==2024.11.0
firebase-admin==6.7.0
flask==3.1.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.88.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-resource-manager==1.14.2
google-cloud-spanner==3.53.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.10.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grapheme==0.6.0
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mesop==1.0.1
msgpack==1.1.0
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pydantic==2.11.3
pydantic-core==2.33.1
pyjwt==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rsa==4.9
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
websockets==15.0.1
werkzeug==3.1.3
